requirements.txt
openai
langchain
chromadb
sentence-transformers
pandas
numpy
streamlit
requests
python-dotenv
langchain_community
geopandas 
folium
langgraph
tqdm
fastapi
uvicorn
scikit-learn

